requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
greenlet==1.1.0
idna==2.10
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.14
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
